\name{RandProd-internal}
\alias{NaReplace}
\alias{NaReplace2}
\alias{NewdataCom}
\alias{Newdata}
\alias{OriginxyCall}
\alias{OriginxyCall2}
\alias{RankComp}
\alias{RankProd1}
\alias{RankProd2}
\alias{RPV1}
\alias{RPV2}
\alias{RPadvanceV2}
\alias{RPadvanceV1}
\alias{RankProd1v2}
\alias{RankSum}
\alias{xyCall}
\title{Internal RankProd functions}
\description{Internal RankProd functions.}
\details{These functions are not meant to be called 
          directly by the user.}
\author{Fangxin Hong, \email{fhong@salk.edu}}
\keyword{internal}
%RankRrod-internal


