\name{RandProd-internal}
\alias{computeOrirank}
\alias{dice.sum.cdf}
\alias{dice.sum.pdf}
\alias{genenames}
\alias{mzmatch.R.wrongvalues}
\alias{NAreplace}
\alias{OriginxyCall}
\alias{plotPFP}
\alias{rankprodbounds}
\alias{xyCall}
\title{Internal RankProd functions}
\description{Internal RankProd functions.}
\details{These functions are not meant to be called 
directly by the user.}
\author{Fangxin Hong, \email{fhong@salk.edu}
\\ Rainer Breitling, \email{rainer.breitling@manchester.ac.uk}
\\ Francesco Del
Carratore, \email{francesco.delcarratore@postgrad.manchester.ac.uk}
\\ Andris Janckevics, \email{andris.jankevics@gmail.com}
}
\keyword{internal}

