\name{RandProd-internal}
\alias{NaReplace}
\alias{NaReplace2}
\alias{NewdataCom}
\alias{Newdata}
\alias{OriginxyCall}
\alias{OriginxyCall2}
\alias{RankComp}
\alias{RankProd1}
\alias{RankProd2}
\alias{RankSum}
\alias{xyCall}
\title{Internal RandProd functions}
\description{Internal RandProd functions.}
\details{These functions are not meant to be called 
          directly by the user.}
\author{Fangxin Hong, \email{fhong@salk.edu}}
\keyword{internal}
%RankRrod-internal


